+++ ruhuna-schedule-ease/app/Http/Controllers/EventController.php
@@ -5,20 +5,21 @@ namespace App\Http\Controllers;
 use App\Models\Event;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
-use Illuminate\Support\Carbon;
-use Illuminate\Support\Facades\Auth;
 
 class EventController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        $this->authorize('read_event', $request->user());
+
         $allevents = Event::all();
         return Inertia::render('Events/EventCalendar', ['allevents' => $allevents]);
-        
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create_event', $request->user());
+
         $data = $request->validate([
             'event_title' => 'required',
             'location' => 'required',
@@ -33,6 +34,8 @@ class EventController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('update_event', $request->user());
+
         $event = Event::findOrFail($id);
 
         $data = $request->validate([
@@ -42,14 +45,15 @@ class EventController extends Controller
             'end' => 'required|date',
         ]);
 
-        
         $event->update($data);
 
         return redirect()->back()->with('success', 'Event updated successfully');
     }
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        $this->authorize('delete_event', $request->user());
+
         try {
             $event = Event::findOrFail($id);
             $event->delete();
@@ -58,4 +62,6 @@ class EventController extends Controller
             return response()->json(['error' => 'Failed to delete event.'], 500);
         }
     }
+
+   
 }
